Rename parameter fullCommand to argumentString
The name fullCommand is misleading and may confuse readers, as it does
not refer to the full command, but only the argument part of the
command.

The parameter has been renamed to argumentString to better reflect
the variable it is supposed to hold, which is the argument part of
the command string.