Refactor isExit method within Command
isExit is currently an abstract method of Command, which requires all
subclasses to implement it. This results in unwanted code duplication
as it has similar behavior amongst all subclasses.

isExit has been changed to a normal method of Command, which returns
false. Hence, only the ByeCommand would need to override the isExit
method in order to return true. isExit methods in the other Command
subclasses have been removed.

This is done to reduce code duplication when adding new commands, and
also make the code easily changeable when we need to change the
implementation of the isExit method.