Add set priority argument on AddCommand
The command uses TaskList.addTodo to add a Todo to the task list.
However, this is insufficient for our needs as we may need to set the
attributes of the Todo before adding them to the task list. By calling
TaskList methods, we print extra output to the console.

We use the new TaskList.addTask method to add tasks/todo to the task
list instead.

By doing this, we can set attributes of the Todo before adding them
to the task list and prevent printing extra lines.