Add saving of Task dates
Saving of Task string to file only inserts a divider when optional
arguments such as category are present. There is difficulty in
determining which optional argument a divider divides.

Task.toFile has been modified to insert blank dividers for optional
arguments, and thus the position of the divider corresponds to the
parameters passed to create new Tasks.

This allows optional arguments to be saved and re-parsed into a new
Task when the program is loaded.