Update Command.execute signature
Command.execute takes in a single TaskList, which was sufficient for
passing in TaskLists, but there needs to be more parameters to add
for new type of ItemList.

The signature of execute has been updated to pass a EnumMap so that
multiple lists can be accessed from the same variable.

This is done so that additional lists can be added to the collection
and passed to the execute function without renaming and adding
parameters in all Command.execute functions.