Refactor Model class
The Command takes in an EnumMap that is created and passed around by
Duke. This is not very OOP as we cannot define our own behaviors for
it. This results in a lot of code duplication in the Duke class, and
many variables being created and passed around as parameters. This makes
refactoring or changing things in the code base very difficult as this
way of coding increases coupling.

A Model class has been refactored and passed into the Command to execute
instead. This also allows us to define its own behaviors such as saving
and loading from the Storage class, and any methods that we would want
to add in the future.

This decouples the Storage class from the Duke class, and is only used
in a single Model class. We can add our own behaviors to the Model class
in the future without adding lots of extra parameters or creating new
objects and passing them around. This also makes the code adhere better
to OOP and SLAP principles.